xdr-codec: restructure u8 packing/unpacking to avoid mistakes
Implementing Pack/Unpack for u8 is a mistake, because they do not
support the normal XDR 4-byte padding rules, which makes it too easy
to accidentally generate bad output. Instead, explicitly support
operations for opaque data, including an Opaque wrapper for [u8].

This is a breaking change; version bumped accordingly.